Fix compile crash when using CALL in #emit
This fixes a crash when calling a function with the CALL instruction in #emit.

Instead of merely writing a symbol's address to the .asm file we check
if we are in a CALL instruction, and if so we output the function's name
rather than its address.

See 7) here: http://forum.sa-mp.com/showthread.php?t=355877

--------- test code --------

foo() {}

main() {
	#emit call foo
}

----- end of test code -----