Fix freeze when include file is a directory
This adds an extra check in plungequalifiedfile() that ensures that
the file being includes is not a directory. It seems to be the only
place where we need to check that, in other places files are opened
by exact name.

The bug appears to be specific to the Linux compiler. On Windows
fopen() will return NULL if trying to open a directory.

Fixes #41.