Don't change __Pawn in compat mode
__Pawn is now fixed at 0x030A, i.e. the new value that was used in
non-compat mode.

I've  come to realize that setting it to the old value doesn't really
make much sense - we still allow the use of the new features but at
the same time pretend to be the old compiler which of course doesn't
have those features.

Maybe we should add another built-in constant for that, perhaps call
it __compat or something similar and define it only when running in
compat mode. The users could then check if the compiler is in compat
mode like this:

  #if defined __compat
      ...
  #endif