Fix symbol nesting level not remembered for 2d+ arrays
This fixes a bug where defining a 2d+ array variable in different
blocks at the same nesting level would trigger a symbol redefinition
error. For example:

main() {
	new x;

	if (x) {
		new a[10][11];
	} else {
		new a[10][12]; // error 021: symbol already defined: "a"
	}
}

It turned out that the compiler defines a separate symbol for each
of the array dimensions but it only was setting sym->compound for the
first of them, causing delete_symbols() to not delete the remaining
dimension symbols. Now it sets the compound field for all dimensions.

Fixes #60