fixes false warning (226) triggers for arrays
Fixes Issue #78

Creates a member variable called 'value_modified' in the value struct
which keeps track if the symbol was modified.

clear_value function sets the member variable to FALSE by default.

When a !,~ or - operation is performed on the symbol, the member is set
to TRUE.

The self-assignment warning is not triggered for arrays unless the
value_modified variable is not set.