fixes inconsistencies with the tagof operator
The standalone tagof operator tries to export tall tags which it works with.

The tagof operator when used as a default argument checks if the tag is zero before exporting.

As the zero tag identifier cannot be exported (even if its `PUBLICTAG` bit is set), it makes sense to use `0` as tag identifier instead of `0 | PUBLICTAG`.

This commit changes the standalone tagof operator behaviour so that the zero tag identifier does not get its `PUBLICTAG` bit set.