corrects behaviour of md-array destructor calls
Currently, the destructors are triggered for each indirection level of the multi-dimensional array. The PAWN Implementer Guide states that the destructor must be triggered for the actual data only.

Apart from the above correction, this commit also fixes a typo which caused wrong sizes to be passed to the destructor.