mark array arguments of public functions modified
The arguments of public functions are marked as READ by default to prevent 203/204 warnings.

The reason is probably that there is no requirement imposed by the host on the user that they read all the arguments.

Smilarly the array/reference arguments need not be modified but the forward takes both cases into consideration.

This prevents warning 214 on array arguments of public functions.